Move wotr types to their own file
